requirements.txt
playwright
python-dotenv
aiohttp
